Replace speech recognition with WhisperKit
